requirements.txt
asgiref==3.2.10
Django==3.1.2
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-js-asset==1.2.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
